refactor(studio.giselles.ai): consolidate invitation acceptance in server actions
