+++ apps/studio.giselles.ai/app/(auth)/join/[token]/login/actions.ts
@@ -1,7 +1,10 @@
 "use server";
 
 import { createClient } from "@/lib/supabase";
+import { captureException } from "@sentry/nextjs";
 import { redirect } from "next/navigation";
+import { JoinError } from "../errors";
+import { acceptInvitation } from "../invitation";
 
 export async function loginUser(formData: FormData) {
 	const email = formData.get("email") as string;
@@ -12,6 +15,16 @@ export async function loginUser(formData: FormData) {
 	if (error) {
 		return { error: error.message };
 	}
-	// If success, redirect to the join page
-	redirect(`/join/${token}`);
+
+	// After successful login, automatically join the team
+	try {
+		await acceptInvitation(token);
+	} catch (err: unknown) {
+		if (err instanceof JoinError) {
+			redirect(`/join/${encodeURIComponent(token)}`);
+		}
+		captureException(err);
+		redirect(`/join/${encodeURIComponent(token)}`);
+	}
+	redirect("/join/success");
 }
